Further optimised by ensuring only NF gizmos are searched.
Previously, we used a set to keep track of NF gizmos searched, however we now ensure gizmos are in normal form on the fly, skipping loops where the gizmos are not in normal form.

Note also that the invariant for Gizmos always being in normal form is removed as a result of this.